+++ UnityFeatureModule/Assets/GameModule/TimeMarker/Scripts/TimeMarkData.cs
@@ -8,9 +8,11 @@
     public class TimeMarkData : ILocalData, IFeatureLocalData
     {
         public Dictionary<string, DateTime> TimeMarks = new Dictionary<string, DateTime>();
+        public Dictionary<string, DateTime> FutureMarks = new Dictionary<string, DateTime>();
         public void Init()
         {
-            this.TimeMarks = new();
+            this.TimeMarks   = new();
+            this.FutureMarks = new();
         }
 
         public Type ControllerType => typeof(TimeMarkDataController);
@@ -23,7 +25,7 @@
         {
             this.localData = localData;
         }
-        
+        #region TimeMark
         public void AddTimeMark(string key, DateTime time)
         {
             var data = (TimeMarkData) this.localData;
@@ -57,7 +59,6 @@
                 return false;
             }
         }
-
         
         public void RemoveTimeMark(string key)
         {
@@ -73,5 +74,58 @@
                 data.TimeMarks[key] = time;
             }
         }
+
+        #endregion
+        #region Future Mark
+
+        public void CreateFutureMark(string key, DateTime futureTime)
+        {
+            var data = this.localData;
+
+            if (data.FutureMarks.ContainsKey(key))
+            {
+                throw new Exception($"TimeMark - Create: Future Mark is already created");
+            }
+            else
+            {
+                data.FutureMarks.Add(key, futureTime);
+            }
+        }
+
+        public DateTime GetFutureMark(string key)
+        {
+            var data = (TimeMarkData) this.localData;
+
+            if (!data.TimeMarks.TryGetValue(key, out var mark))
+            {
+                throw new Exception($"TimeMark - Get: Future Mark have not been created");
+            }
+            return mark;
+        }
+
+        public void RemoveFutureMark(string key)
+        {
+            var data = (TimeMarkData)this.localData;
+            if (!data.FutureMarks.ContainsKey(key))
+            {
+                throw new Exception($"TimeMark - Remove: Future Mark have not been created");
+            }
+            data.FutureMarks.Remove(key);
+        }
+
+        public void UpdateFutureMark(string key, DateTime time)
+        {
+            var data = (TimeMarkData)this.localData;
+            if (data.FutureMarks.ContainsKey(key))
+            {
+                data.FutureMarks[key] = time;
+            }
+            else
+            {
+                throw new Exception($"TimeMark - Update: Future Mark have not been created");
+            }
+        }
+
+        #endregion
     }
 }